% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowGate-package.R
\docType{package}
\name{flowGate-package}
\alias{flowGate}
\alias{flowGate-package}
\title{flowGate: Interactive Cytometry Gating in R}
\description{
To learn more about how to use flowGate, please start by browsing the vignettes:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.bioconductor.org/packages/release/bioc/html/flowGate.html}
  \item Report bugs at \url{https://github.com/NKInstinct/flowGate/issues}
}

}
\author{
\strong{Maintainer}: Andrew Wight \email{andrew.wight10@gmail.com} (\href{https://orcid.org/0000-0003-3116-8722}{ORCID})

Authors:
\itemize{
  \item Harvey Cantor \email{Harvey_Cantor@dfci.harvard.edu} (\href{https://orcid.org/0000-0002-3313-2478}{ORCID}) [laboratory director]
}

}
\keyword{internal}
