% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_apply_gate_check.R
\name{gs_apply_gate_check}
\alias{gs_apply_gate_check}
\title{The purpose of this function is to allow the readjustment of a gate on an
individual specimen via the Shiny app, quickly iterating through every
specimen present in the GatingSet.}
\usage{
gs_apply_gate_check(gs, gate, sample = NULL, ...)
}
\arguments{
\item{gs}{A GatingSet or list of GatingSets.}

\item{gate}{A tibble-formatted gating strategy (see examples below)}

\item{sample}{Sample index in the GatingSet, default NULL will iterate through
all specimens in the GatingSet.}

\item{...}{Other parameters to pass to gs_gate_interactive(). Note that only
constant parameters should be supplied here---anything that varies should
be included in the gating_strategy tibble.}
}
\value{
the GatingSet or list of GatingSets with the gates in gating_strategy
  applied as specified.
}
\description{
This allows for fine-tuning of the gate placement for non-representative
specimens, or when automated gating methods fail.
}
\examples{

fs <- flowCore::read.flowSet(
  path = system.file("extdata", package = "flowGate"), pattern = ".FCS$")

gs <- flowWorkspace::GatingSet(fs)

# Note - this is a very rudamentary GatingSet for example purposes only. 
# Please see the vignette accompanying this package or the flowWorkspace
# documentation # for a complete look at creating a GatingSet.

gating_strategy <- tibble::tribble(
~filterId,      ~dims,                       ~subset,        ~coords,
"Lymphocytes", list("FSC-H", "SSC-H"),       "root",         list(c(0, 3e5), c(0, 3e5)),
"CD45 CD15",   list("CD45 PE", "CD15 FITC"), "Lymphocytes",  list(c(0, 3e5), c(0, 2e5)),
)


if(interactive()){
gs_apply_gating_strategy(gs,
gating_strategy = gating_strategy,
bins = 512) # note that extra args for gs_gate_interactive can be supplied.
}
}
