% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_to_openCyto.R
\name{gs_to_openCyto}
\alias{gs_to_openCyto}
\title{Converts an existing GatingSet object into an openCyto template,
based off gates existing for given sample. Experimental}
\usage{
gs_to_openCyto(gs, sample = 1, outpath = NULL, filename = "openCytoTemplate")
}
\arguments{
\item{gs}{A GatingSet object}

\item{sample}{The sample index to use, default is 1}

\item{outpath}{The file.path to the storage location. Default NULL will
result in creation within the current working directory}
}
\value{
A .csv file containing an openCyto style template for
subsequent re-gating attempts
}
\description{
Converts an existing GatingSet object into an openCyto template,
based off gates existing for given sample. Experimental
}
\examples{
A <- 2 + 2

}
